+++ kiltergpt/data/datamodules.py
@@ -17,6 +17,8 @@ class KilterDataModule(L.LightningDataModule):
         pin_memory: bool = True,
         context_len: int = 64,
         label_smoothing: bool = True,
+        prompt_size: float = 0.2,
+        min_tokens: int = 5,
     ):
         super().__init__()
         if not isinstance(data_dir, Path):
@@ -27,6 +29,8 @@ class KilterDataModule(L.LightningDataModule):
         self.pin_memory = pin_memory
         self.context_len = context_len
         self.label_smoothing = label_smoothing
+        self.prompt_size = prompt_size
+        self.min_tokens = min_tokens
 
     def setup(self, stage=None):
         self.tokenizer = Tokenizer()
@@ -35,6 +39,8 @@ class KilterDataModule(L.LightningDataModule):
             self.tokenizer,
             context_len=self.context_len,
             label_smoothing=self.label_smoothing,
+            prompt_size=self.prompt_size,
+            min_tokens=self.min_tokens,
         )
 
         self.val = KilterGPTDataset(
@@ -42,19 +48,27 @@ class KilterDataModule(L.LightningDataModule):
             self.tokenizer,
             context_len=self.context_len,
             label_smoothing=self.label_smoothing,
+            prompt_size=self.prompt_size,
+            min_tokens=self.min_tokens,
         )
         self.test = KilterGPTDataset(
             self.data_dir / "test.csv",
             self.tokenizer,
             context_len=self.context_len,
             label_smoothing=self.label_smoothing,
+            prompt_size=self.prompt_size,
+            min_tokens=self.min_tokens,
         )
         self.test.eval = True
         self.vocab_size = self.tokenizer.vocab_size
 
     def _get_dataloader(self, dataset, shuffle: bool = False) -> DataLoader:
         return DataLoader(
-            dataset, batch_size=self.batch_size, shuffle=shuffle, pin_memory=True, num_workers=self.num_workers
+            dataset,
+            batch_size=self.batch_size,
+            shuffle=shuffle,
+            pin_memory=True,
+            num_workers=self.num_workers,
         )
 
     def train_dataloader(self) -> DataLoader: