+++ kiltergpt/models/gpt.py
@@ -7,7 +7,8 @@ import torch.nn.functional as F
 import torchmetrics.functional as M
 from fastapi import FastAPI
 
-from ..utils import Plotter, WarmupCosineSchedule, get_histogram
+from ..utils import Plotter, WarmupCosineSchedule
+from ..utils.metrics import get_histogram, jaccard_similarity
 
 
 class LayerNorm(nn.Module):
@@ -186,11 +187,23 @@ class GPTModel(L.LightningModule):
             self.test_generated.append(generated)
             self.test_real.append(target)
 
-    def on_test_epoch_end(self):
+    def _get_hist_pearson(self):
+        """Calculate the spearman correlation between token distributions of real and genenerated sequences"""
         hist_generated = get_histogram(self.test_generated, self.tokenizer.vocab_size)
         hist_real = get_histogram(self.test_real, self.tokenizer.vocab_size)
         hist_pearson = M.spearman_corrcoef(hist_generated, hist_real)
-        self.log("test/hist_pearson", hist_pearson)
+        return hist_pearson
+
+    def _get_jaccard_similarity(self):
+        generated_onehot = torch.stack([self.tokenizer.onehot(x) for x in self.test_generated])
+        real_onehot = torch.stack([self.tokenizer.onehot(x) for x in self.test_real])
+        jaccard = jaccard_similarity(generated_onehot, real_onehot)
+        return jaccard
+
+    def on_test_epoch_end(self):
+        hist_pearson = self._get_hist_pearson()
+        jaccard_similarity = self._get_jaccard_similarity()
+        self.log_dict({"test/hist_pearson": hist_pearson, "test/jaccard_similarity": jaccard_similarity.mean()})
 
     def configure_optimizers(self):
         param_dict = {pn: p for pn, p in self.named_parameters()}